Eliminate special case for URL navigation on the states
Rationale:
After giving this some thought it seems like the URL, or the navigation
to an URL, should be defined as a transition and not a state. Unlike the
remaining things that define the state, the URL is actually imposing a
mutation on the browser (and such things are normally defined in
transitions).

Changes:
* Force evaluation of the state predicates.
* Fix bug on `value?` predicate that was using the Taxi `visible?`
  function inappropriately.
* Change `run-route` to expect a transition in the beginning.
* Change `run-transition` to first run the transitions and only then
  check the state.