Change route traversal strategy and rollback graph representation
Motivation:
The changes to the graph representation introduced in `3fd7437` do not
make sense. To have a complete representation of the graph, we need both
the nodes and the edges.

Instead of having a transition, we now have the URL as a parameter to the
`run` function and we navigate to that URL before starting the traversal
of the route.